portal26_otel_agent/requirements.txt
google-adk>=1.17.0
litellm>=1.75.5
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-google-genai>=0.4b0
opentelemetry-instrumentation-vertexai>=2.0b0
python-dotenv
flask
gunicorn
boto3
